resheduler: delete sts for stable reconciliation
