+++ modules/500-upmeter/hooks/smokemini/internal/scheduler/node_by_zone_test.go
@@ -20,8 +20,6 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
-
 	"github.com/deckhouse/deckhouse/modules/500-upmeter/hooks/smokemini/internal/snapshot"
 )
 
@@ -129,61 +127,6 @@ func Test_spread(t *testing.T) {
 			t.Errorf("spread should not modify input")
 		}
 	}
-}
-
-func Test_nodeStateFilterPipe_Filter(t *testing.T) {
-	type args struct {
-		nodes []snapshot.Node
-		x     string
-	}
-
-	byZone := &filterBy{func(node snapshot.Node) bool {
-		return node.Zone == "z"
-	}}
-
-	bySchedulable := &filterBy{func(node snapshot.Node) bool {
-		return node.Schedulable
-	}}
-
-	tests := []struct {
-		name string
-		pipe NodeFilterPipe
-		args args
-		want []snapshot.Node
-	}{
-		{
-			name: "no nodes return no nodes",
-			pipe: NodeFilterPipe{byZone, bySchedulable},
-			args: args{nodes: []snapshot.Node{}, x: ""},
-			want: []snapshot.Node{},
-		},
-		{
-			name: "filters overlap ",
-			pipe: NodeFilterPipe{byZone, bySchedulable},
-			args: args{nodes: []snapshot.Node{
-				{Zone: "_", Schedulable: true},
-				{Zone: "z", Schedulable: true},
-				{Zone: "z", Schedulable: false},
-			}},
-			want: []snapshot.Node{{Zone: "z", Schedulable: true}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := tt.pipe.Filter(tt.args.nodes, tt.args.x)
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("NodeFilterPipe.Filter() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-type filterBy struct {
-	filter func(snapshot.Node) bool
-}
-
-func (f *filterBy) Filter(nodes []snapshot.Node, _ string) []snapshot.Node {
-	return applyFilter(nodes, f.filter)
 }
 
 func stateByZones(xs ...string) State {
@@ -316,130 +259,6 @@ func Test_filterByZone_selectZone(t *testing.T) {
 			if got := f.selectZone(tt.args.nodes, tt.args.x); got != tt.want {
 				t.Errorf("filterByZone.selectZone() = %v, want %v", got, tt.want)
 			}
-		})
-	}
-}
-
-func Test_FilterByMinSts_Filter(t *testing.T) {
-	// The filter by sts has nothing to do with zones. It works with nodes it gets, and makes decisions
-	// based on sts counts.
-	type args struct {
-		nodes []snapshot.Node
-		x     string
-	}
-
-	tests := []struct {
-		name  string
-		state func() State
-		args  args
-		want  []snapshot.Node
-	}{
-		{
-			name:  "no input nodes results no output nodes",
-			state: fakeState,
-			args:  args{nodes: []snapshot.Node{}, x: "c"},
-			want:  []snapshot.Node{},
-		},
-		{
-			name: "one node for all leaves the node as is",
-			state: func() State {
-				state := fakeState()
-				node := fakeNode(1)
-				for _, sts := range state {
-					sts.Node = node.Name
-					sts.Zone = node.Zone
-				}
-				return state
-			},
-			args: args{nodes: []snapshot.Node{fakeNode(1)}, x: "c"},
-			want: []snapshot.Node{fakeNode(1)},
-		},
-		{
-			name: "when all 2 nodes occupied, current sts node is omitted",
-			state: func() State {
-				state := fakeState()
-				indexes := []string{"a", "b", "c", "d", "e"}
-				for i, x := range indexes {
-					node := fakeNode(i%2 + 1)
-					state[x].Node = node.Name
-					state[x].Zone = node.Zone
-				}
-				return state
-			},
-			args: args{
-				x: "c",
-				nodes: []snapshot.Node{
-					fakeNode(1), // a, c, e
-					fakeNode(2), // b, d
-				},
-			},
-			// exclude node with c
-			want: []snapshot.Node{fakeNode(2)},
-		},
-		{
-			name: "when all 5 nodes occupied, current sts node is omitted",
-			state: func() State {
-				state := fakeState()
-				indexes := []string{"a", "b", "c", "d", "e"}
-				for i, x := range indexes {
-					node := fakeNode(i + 1)
-					state[x].Node = node.Name
-					state[x].Zone = node.Zone
-				}
-				return state
-			},
-			args: args{
-				x: "c",
-				nodes: []snapshot.Node{
-					fakeNode(1), // a
-					fakeNode(2), // b
-					fakeNode(3), // c
-					fakeNode(4), // d
-					fakeNode(5), // e
-				},
-			},
-			want: []snapshot.Node{
-				// exclude c
-				fakeNode(1), // a
-				fakeNode(2), // b
-				fakeNode(4), // d
-				fakeNode(5), // e
-			},
-		},
-		{
-			name: "when all 4/5 nodes occupied, only free node is returned",
-			state: func() State {
-				state := fakeState()
-				indexes := []string{"a", "b", "c", "d", "e"}
-				for i, x := range indexes {
-					node := fakeNode(i + 1)
-					state[x].Node = node.Name
-					state[x].Zone = node.Zone
-				}
-				state["c"] = state["b"]
-				return state
-			},
-			args: args{
-				x: "c",
-				nodes: []snapshot.Node{
-					fakeNode(1), // a
-					fakeNode(2), // b, c
-					fakeNode(3), // —
-					fakeNode(4), // d
-					fakeNode(5), // e
-				},
-			},
-			want: []snapshot.Node{fakeNode(3)},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			f := &filterByMinSts{state: tt.state()}
-
-			got := f.Filter(tt.args.nodes, tt.args.x)
-
-			assert.Equal(t, tt.want, got)
 		})
 	}
 }
