+++ modules/500-upmeter/hooks/smokemini/internal/scheduler/statefulset_by_storageclass.go
@@ -1,0 +1,31 @@
+/*
+Copyright 2021 Flant JSC
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package scheduler
+
+type selectByStorageClass struct {
+	storageClass string
+}
+
+func (s *selectByStorageClass) Select(state State) (string, error) {
+	// Find sts with outdated storage class
+	for x, sts := range state {
+		if sts.StorageClass != s.storageClass {
+			return x, nil
+		}
+	}
+	return "", errNext
+}
